pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "prompt-toolkit>=3.0.0",
]

[project.scripts]
agent = "main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]
packages   = ["core", "cli", "tools", "llm", "memory"]
